pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neoantigen-vaccine-pipeline"
version = "0.1.0"
description = "Personalized mRNA neoantigen vaccine design pipeline"
readme = "plan.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "McClain Thiel"}]

dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "biopython>=1.81",
    "scipy>=1.11",
    "pyyaml>=6.0",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "httpx>=0.24",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
